revert: restore Filament Actions on ModerationQueue Livewire component
Revert the Alpine dispatch approach. Actions (takeAction, escalate,
dismiss) are back on the Livewire component using InteractsWithActions
+ InteractsWithForms traits. Rendered via {{ $this->takeActionAction }}
in blade with <x-filament-actions::modals /> in the component view.
