+++ app-modules/panel-admin/resources/views/moderation/queue.blade.php
@@ -1,5 +1,3 @@
-<x-filament-panels::page
-    x-on:mount-queue-action.window="$wire.mountAction($event.detail.action, $event.detail.arguments ?? {})"
->
+<x-filament-panels::page>
     <livewire:moderation-queue />
 </x-filament-panels::page>
