+++ app-modules/panel-admin/resources/views/moderation/queue/case-detail.blade.php
@@ -299,8 +299,7 @@
             {{-- Action buttons --}}
             @if ($case->status->isOpen())
                 <section
-                    x-data
-                    class="sticky top-4 rounded-lg border border-zinc-200/80 bg-white p-4 dark:border-white/5 dark:bg-white/[0.03]"
+                    class="sticky top-4 rounded-lg border border-zinc-200/80 bg-white p-3 sm:p-4 dark:border-white/5 dark:bg-white/[0.03]"
                 >
                     <h3
                         class="mb-3 flex items-center gap-2 text-[11px] font-semibold tracking-wider text-zinc-400 uppercase dark:text-zinc-500"
@@ -312,35 +311,8 @@
                             )
                         }}
                     </h3>
-                    <div class="flex flex-col gap-2">
-                        <button
-                            type="button"
-                            x-on:click="$dispatch('mount-queue-action', { action: 'takeAction', arguments: { caseId: '{{ $case->id }}' } })"
-                            class="bg-danger-600 hover:bg-danger-500 dark:bg-danger-500 dark:hover:bg-danger-400 inline-flex w-full items-center justify-center gap-2 rounded-lg px-4 py-2.5 text-sm font-semibold text-white shadow-sm transition"
-                        >
-                            <x-heroicon-o-bolt class="size-4" />
-                            {{
-                                __(
-                                    'panel-admin::moderation.queue.actions.take_action',
-                                )
-                            }}
-                        </button>
-                        <button
-                            type="button"
-                            x-on:click="$dispatch('mount-queue-action', { action: 'escalate', arguments: { caseId: '{{ $case->id }}' } })"
-                            class="bg-warning-500 hover:bg-warning-400 dark:bg-warning-600 dark:hover:bg-warning-500 inline-flex w-full items-center justify-center gap-2 rounded-lg px-4 py-2.5 text-sm font-semibold text-white shadow-sm transition"
-                        >
-                            <x-heroicon-o-arrow-up-circle class="size-4" />
-                            {{ __('panel-admin::moderation.queue.actions.escalate') }}
-                        </button>
-                        <button
-                            type="button"
-                            x-on:click="$dispatch('mount-queue-action', { action: 'dismiss', arguments: { caseId: '{{ $case->id }}' } })"
-                            class="inline-flex w-full items-center justify-center gap-2 rounded-lg border border-zinc-300 bg-white px-4 py-2.5 text-sm font-semibold text-zinc-700 shadow-sm transition hover:bg-zinc-50 dark:border-white/10 dark:bg-white/5 dark:text-zinc-300 dark:hover:bg-white/10"
-                        >
-                            <x-heroicon-o-x-circle class="size-4" />
-                            {{ __('panel-admin::moderation.queue.actions.dismiss') }}
-                        </button>
+                    <div class="[&_.fi-ac-action]:w-full [&_.fi-ac-action]:justify-center flex flex-col gap-2">
+                        {{ $this->takeActionAction }} {{ $this->escalateAction }} {{ $this->dismissAction }}
                     </div>
                 </section>
             @endif
