+++ app-modules/panel-admin/resources/views/moderation/queue/index.blade.php
@@ -152,4 +152,6 @@
             @endif
         </div>
     </div>
+
+    <x-filament-actions::modals />
 </div>
